+++ src/main/java/rocks/cleancode/hamcrest/file/FileMatchers.java
@@ -1,5 +1,7 @@
 package rocks.cleancode.hamcrest.file;
 
+import org.hamcrest.Matcher;
+
 import java.io.File;
 
 /**
@@ -16,8 +18,19 @@ public class FileMatchers {
      *
      * @since 1.0.0
      */
-    public static IsFileMatcher<File> file() {
+    public static Matcher<File> file() {
         return new IsFileMatcher<>(File::isFile);
+    }
+
+    /**
+     * Create a matcher for existing directory.
+     *
+     * @return Matcher for existing directory.
+     *
+     * @since 1.0.0
+     */
+    public static Matcher<File> directory() {
+        return new IsDirectoryMatcher<>(File::isDirectory);
     }
 
     private FileMatchers() {