+++ src/test/java/rocks/cleancode/hamcrest/file/PathMatchersTest.java
@@ -10,6 +10,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static rocks.cleancode.hamcrest.file.PathMatchers.directory;
 import static rocks.cleancode.hamcrest.file.PathMatchers.file;
 
 public class PathMatchersTest {
@@ -34,6 +35,31 @@ public class PathMatchersTest {
             "%n%s%n%s",
             "Expected: is a file",
             "     but: was not a file"
+        );
+
+        assertThat(assertionError.getMessage(), is(equalTo(expectedMessage)));
+    }
+
+    @Test
+    public void should_match_existing_directory() throws URISyntaxException {
+        Path existingDirectory = Paths.get(getClass().getResource("/test_directory").toURI());
+
+        assertThat(existingDirectory, is(directory()));
+    }
+
+    @Test
+    public void should_fail_when_directory_does_not_exist() {
+        Path notADirectory = Paths.get("/not/a/directory");
+
+        AssertionError assertionError = assertThrows(
+            AssertionError.class,
+            () -> assertThat(notADirectory, is(directory()))
+        );
+
+        String expectedMessage = String.format(
+            "%n%s%n%s",
+            "Expected: is a directory",
+            "     but: was not a directory"
         );
 
         assertThat(assertionError.getMessage(), is(equalTo(expectedMessage)));