+++ CHANGELOG.md
@@ -4,6 +4,6 @@
 
 ### Features
 
-- Matcher to validate existing file
+- Matchers to validate existing file and directory
 
 [Unreleased]: https://github.com/clean-code-rocks/hamcrest-java-file/commits/main